requirements.txt
PyQt6>=6.4.0
opencv-python>=4.8.0
numpy>=1.24.0
Pillow>=10.0.0
python-dotenv>=1.0.0
openai>=1.3.0
pynput>=1.7.6
winotify
win10toast>=0.9.0
plyer>=2.1.0
pytesseract>=0.3.10